refactor: :recycle: organize math utilities for clarity
