+++ src/main.ts
@@ -1,5 +1,5 @@
 import { VisualizationScene } from './visualization';
-import { Vector3, Matrix3x3 } from './mathTypes';
+import { Vector3, Matrix3x3, multiplyMatrices, multiplyMatrixVector } from './mathTypes';
 
 // State variables
 let rawScene: VisualizationScene | null = null;
@@ -40,21 +40,6 @@ function normalizeHeadingRange(heading: number): number {
         heading -= 360;
     }
     return heading;
-}
-
-// Dot product: takes two vectors and returns the sum of component-wise multiplication
-function dotProduct(v1: Vector3, v2: Vector3): number {
-    return v1[0] * v2[0] + v1[1] * v2[1] + v1[2] * v2[2];
-}
-
-// Matrix-vector multiplication: result = matrix × vector
-// Each component of the result is the dot product of the corresponding matrix row with the vector
-function multiplyMatrixVector(matrix: Matrix3x3, vector: Vector3): Vector3 {
-    return [
-        dotProduct(matrix[0], vector),
-        dotProduct(matrix[1], vector),
-        dotProduct(matrix[2], vector),
-    ];
 }
 
 // Build the rotation matrix from Euler angles (alpha, beta, gamma)
@@ -159,23 +144,6 @@ function buildRotationMatrixFromAngles(heading: number, elevation: number, roll:
     const R: Matrix3x3 = multiplyMatrices(Rr, ReRh);
 
     return R;
-}
-
-// Matrix multiplication: result = A × B
-function multiplyMatrices(A: Matrix3x3, B: Matrix3x3): Matrix3x3 {
-    const result: Matrix3x3 = [
-        [0, 0, 0],
-        [0, 0, 0],
-        [0, 0, 0]
-    ];
-
-    for (let i = 0; i < 3; i++) {
-        for (let j = 0; j < 3; j++) {
-            result[i][j] = A[i][0] * B[0][j] + A[i][1] * B[1][j] + A[i][2] * B[2][j];
-        }
-    }
-
-    return result;
 }
 
 // Handle device orientation event
